Update BIGS903.txt
The BIGS903 software applicatio enables the user to input a user-defined subset of game rules. The user-defined subset of game rules specifies the composition of each deck, and the number of decks in the stack. If each deck includes at least one joker, then the user-defined subset of game rules includes a rule selected by the user from a group of two rules wherein: The first of the group of two rules is each and every joker has a value of zero points. The second of the group of two rules is each and every joker has a value of any number of points required to make a hand total of twenty-seven points. Given the user-defined subset of game rules, the BIGS903 software application builds a set of databases of player-finish probabilities. The BIGS903 software application saves the set of databases and data about the user-defined subset of game rules to a set of data files.