Add files via upload
The BIGS907 software application enables the user to do the following. Load into program memory a set of databases of player-finish probabilities saved to a set of data files by the BIGS903 software application. Load into program memory a set of databases of dealer-finish probabilities saved to a set of data files by the BIGS906 software application. Load into program memory data about a user-defined subset of game rules saved to a data file by the BIGS903 software application. Input a user-defined set of game rules consisting of the user-defined subset of game rules and the remaining game rules required to play a double-exposure game. Find the set of combination-dependent-playing-strategy tables for a double-exposure game that is subject to the user-defined set of game rules. Calculate an estimate of the expected value of the game wager for the double-exposure game that is subject to the user-defined set of game rules.