Add files via upload
The BIGS908 software application is a game simulator. The component enables the user to do the following. Load into program memory a user-defined set of game rules saved to a data file by a software application selected from a group consisting of the BIGS907 software application and the BIGS921 software application. Load into program memory a matching-strategy table selected from a group consisting of a set of combination-dependent-playing-strategy tables saved to a data file by the BIGS907 software application and a total-dependent-basic-strategy table saved to a data file by the BIGS921 software application. Use the matching-strategy table to make decisions on how to play the player's hand. Use the predetermined strategy for the play of the dealer's hand specified by the user-defined set of game rules to make decisions on how to play the dealer's hand. Simulate millions of rounds of play of the double exposure game in accordance with the user-defined set of game rules.