Update BIGS921.txt
For each of the combination of card ranks that could be the combination of card ranks assigned to the cards removed from the stack and dealt face up to form the set of hands consisting of the player's initial hand and the dealer's initial hand, the BIGS907 software application does all of the following. The BIGS907 software application performs an abbreviated-combinatorial analysis to calculate the expected value of the game wager for each of the strategies for the play of the player's hand that has at least a chance of being optimal, and saves the results of these calculations to a data file with a 't' character as the file name suffix. The BIGS907 software application calculates the probability of the occurrence of the combination of card ranks and saves the results of these calculations to a data file with a 'p' character as the file name suffix.

The BIGS921 software application uses the data contained in these files to convert the set of combination-dependent-playing-strategy tables into a total-dependent-basic-strategy table. Accordingly, the BIGS921 software application is capable of producing a set of files containing data about the total-dependent-basic-strategy table for a double exposure game.