5cph.BAS
REM "This program creates a set of arrays. Arrays are simply variables"
REM "with a number elements that can each hold data in program memory."
REM "The set of arrays includes an array dimensioned as phv0(8567) and"
REM "an array dimensioned as phv1(8567). The elements in these arrays"
REM "represent poker hands. Each poker hand is made from a set of five"
REM "cards. Each card has a denomination or rank. Each rank is a rank"
REM "selected from group of ranks. The group ranks consists of King,"
REM "Queen, Jack, 10, 9, 8, 7, 6, 5, 4, 3, 2, Ace, and Joker. The 
REM "Program uses the numbers 13, 12, 11, 10, 9, 8, 7, 6, 5, 4, 3, 2,"
REM "1, and 0 to represent the ranks King, Queen, Jack, 10, 9, 8, 7,"
REM "6, 5, 4, 3, 2, Ace and Joker. Accordingly, each card of each poker"
REM "hand is assigned a number that corresponds to the rank of the card."
REM "The program arranges the cards of each poker hand by the numbers"
REM "assigned to each card so that the cards are arranged from the highest"