1sim.BAS
The 1sim software application is a game simulator. The component enables the user to do the following. Load into program memory a user-defined set of game rules saved to data file by the 1roi software application. Load into program memory a matching-basic-strategy table saved to data file by the 1roi software application. Thereby, program a computer-controlled player to use the matching-basic-strategy table to make decisions on how to play the player's hand. Program a computer-controlled dealer to use a matching-predetermined strategy to make decisions on how to play the dealer's hand. Simulate the play of millions of rounds of the single exposure game of any of a subset of optional embodiments subject to the user-defined set of game rules. In a single exposure game, the dealer forms the dealer's-initial hand either by dealing a first card face down and dealing a second card face up, or by dealing a first card face up, proceeding through any additional steps required to complete the player's hand, a