Update 1sim.BAS
The 1sim.BAS software application is run by game simulation program The program is written in the QB64 computer language. The QB64 computer language is a version of the basic computer language. The basic computer language, in case you don't already know, is regarded as one of the easiest computer languages to use. I think I learned it in junior high school. The QB64 computer language, in case you don't already know, compiles to the c++ computer language. QB64 computer language is sort of a supped up version of MicrosoftQB45. The 1sim.BAS software application is run by a game simulation program written in the QB64 computer language complied into a c++ computer language run software application by the QB64 software application complier.

The 1sim.BAS software application keeps track of statistics about the number of times that each value of a group of values occurs during a game simulation of a poker-like Finnish 27 ™ game that is subject to a user-defined set of game rules. I made the following changes to the game simulation program. The program now keeps track of statistics about the number of times that each value of a group of values occurs. The group of values consists of the number of times the player made each hand of a group of hands. The group of hands consists of two groups. The two groups are the 5-card 27 player hands group and 5-card player hands with a numerical value of 26 or less group.   For that reason, now the approximation of the probability of the occurrence of each rank category of 5-card player hands with a numerical sum value of 26 or less can be calculated as the quotient of the number times that the player made 5-card player hands with a numerical sum value of 26 or less,  divided by the total number of times that the player made a hand. The probability of the occurrence of each rank category of 5-card player hand that has a numerical sum value of 27, can be calculated as the quotient of the number of times that the player made 5-card-player hands with a numerical sum value of 27 divided by the number of times that the player made a hand.