Update 1roi.BAS
The 1roi.BAS software application is run by a combinatorial analysis program The program is written in the QB64 computer language. The QB64 computer language is a version of the basic computer language. The basic computer language, in case you don't already know, is regarded as one of the easiest computer languages to use. I think I learned it in junior high school. The QB64 computer language, in case you don't already know, compiles to the c++ computer language. QB64 computer language is sort of a supped up version of MicrosoftQB45. The 1roi.BAS software application is run by a combinatorial analysis program written in the QB64 computer language compiled into a c++ computer language run software application by the QB64 software application complier.

The 1roi.BAS software application finds the best strategy for the play of the player's hands in a poker-like Finnish 27 ™ game that is subject to a user-defined set of game rules. Based on the best strategy for the play of the player's hand, the 1roi.BAS software application calculates the house edge of a poker-like Finnish 27 ™ game that is subject to the user-defined set of game rules. I made the following changes to the combinatorial analysis program. The program now enables to specify that the dealer pays the player a bonus payout whenever the player makes a hand consisting of five cards with a numerical sum value of 27.  For that reason, it now possible to find the best strategy for the play of the player's hand in a poker-like Finnish 27 ™ game wherein; the user specifies that the dealer pays the player a bonus payout whenever the player makes a hand consisting of five cards with a numerical sum value of 27.